[KGEN] When escaping a parameter name, use *"foo" instead of "foo" (#2778)
This makes the grammar for escaped identifiers less ambiguous with
string literals, which makes way for future string support.  We'd
ideally like to support `foo` style identifiers, but the MLIR
parser doesn't support such a thing.

This approach is ugly, but is only needed when a parameter name
conflicts with a keyword, which shouldn't happen in practice.

MODULAR_ORIG_COMMIT_REV_ID: 8096122f2c921573cf6073da54e15fb8a260c461