[MLIRToBEF] Greatly optimize DenseElementsAttr/DenseArrayAttr encoding (#2834)
The current emission is suboptimal for a few reasons, and this commit
rectifies this with two main changes.

Firstly, a new ConcatentationTreeNode is added to support emitting an
ArrayRef of data whose lifetime is guaranteed to live beyond BEF emission (e.g.
all attributes satisfy this). This allows for us to only copy the attribute data when
actually generating the output.

Secondly, the encoding for DenseElementsAttr previously went through
APInt/APFloat ranges, which is horribly inefficient for large data. We rewrite the
DenseElementsAttr emission to make use of the raw data API when possible (which works in
all cases except for bool, which needs to be handled specially).

This gives >100x speedup when compiling a 1gb TFLite model in an asan
build (from minutes to seconds).

MODULAR_ORIG_COMMIT_REV_ID: 08aa7e320cb1148d2140075ef2432246f6484b5b