[KGEN] Add a way to compare objects to `OpaqueObjectInterface`. (#2856)
We have to ensure that the implementations of a given interface are
self-consistent. To this end, we implement
`OpaqueObjectInterface::equal` which allows a user-defined type to
determine equality of 2 objects of that type. The user can effectively
opt-out of the comparison by simply always returning 'true' for the
implementation of `OpaqueObjectInterface::equal`. The user can also
control the comparison by setting their own per-type tolerance, or even
adding it to the tag attribute for the result.

MODULAR_ORIG_COMMIT_REV_ID: 97b5b6c3d848464ef80e1ca96922a1fa40d99722