[KGEN] Introduce a new !kgen.region type. (#2918)
This is used to represent regions (and symbols) that are passed as
parameters.  This includes bindings for both value types as well as
any parameters that a bound generator might take.  This is sugared
in a parameter list to simple function type syntax when there are no
bound parameters.

This introduces the first parameter type that may itself be
parameterized, so this adds checking for this.

This is part of Issue #2883

MODULAR_ORIG_COMMIT_REV_ID: e67f2a1ff8673145953db32cf9290d1b7ae81ff3