[KGEN] Introduce a new `SymbolConstantAttr` class and support logic. (#2920)
This allows binding !kgen.region parameters to a symbol at a kgen.call
site.  This requires a significant amount of type checking to make sure
the referenced symbol matches the declared type.  This is also the first
case where the type of a bound parameter is itself subject to
substitution.

This is part of Issue #2883.

MODULAR_ORIG_COMMIT_REV_ID: a56f0f6f0c9e9cb415d00136728773732fd590b0