[KGEN] Move SymbolConstantAttr verification into ParameterVerifier. (#2925)
KGEN needs to check that a use of a symbol has the correct value
and parameter types specified.  Previously, I special cased this
for kgen.call, but these can potentially exist in nested positions
and with other operations.  As such it is better to perform this
in the ParameterVerifier, which is already doing this stuff.

This required several other refactorings:

1) Give ParameterDeclsAndUses::calculateAndVerify a
   SymbolTableCollection so it can efficiently do the verification.
2) Change all the declarations to do their body verification in
   verifySymbolUses so they have it.
3) Change ParameterVerifier to have a notion of current operation
   (so it can generate good diagnostics on a problem), which
   eliminates having to pass down a Location in collection pass.

MODULAR_ORIG_COMMIT_REV_ID: 5a99b24f1d9cc311f0cd8becbdeda1f0c3af3011