[KGEN] Introduce a new `kgen.call_param` op. (#2927)
This operation models a call to a parameter that is bound to a
symbol.  This patch implements enough to round trip it and to
canonicalize it away when directly invoking an symbol, but
the elaborator still needs work.

MODULAR_ORIG_COMMIT_REV_ID: f64c22d9cdbc8942f19c380f6d6a4012f768f0c7