[KGEN] Rearrange the KGEN impl details into a new KGENUtils file. (#2937)
This moves a bunch of the generic printing and parsing logic out
of KGENOps and KGENAttrs, allowing the former to focus more on
implementation detail of operations and the later to focus on the
implementation details of attributes.  The printing/parsing logic
is shared across both and as well as types, so we had a bit of an
accidental divide going on.

MODULAR_ORIG_COMMIT_REV_ID: 1ad7b52289013eec731ad7b28bd772d9cd9ca363