[KGEN] Diagnose incorrect signature types modularly. (#2945)
This rejects an signature type that uses an undefined input parameter
or uses an input parameter with the wrong type.

Much of this is refactoring the existing ParameterVerifier into a
ParameterCollector base class that can be used for this as well as
fixing all the problems in the testsuite that this now correctly
rejects.

MODULAR_ORIG_COMMIT_REV_ID: e894b38b9770205b76400b4c3ef5f95834c2892f