[KGEN] Add basic target info attribute. (#3006)
This patch adds an attribute that we can use for propagating target
information to the various target-specific ops that are coming in future
patches. The attribute is typed so that it can be used as a parameter to
a generator/interface.

The attribute is expected to evolve in the future, it absolutely should
not simply stay as a target triple, but this at least allows us to refer
to the notion of a target.

MODULAR_ORIG_COMMIT_REV_ID: 79a1f028b8db4ecabf1f6fc67dc707100a7bc3f2