[BEF] Add encoding/decoding support for resource arrays (#2983)
Resources are used to hold large arrays of data, as they do
not get allocated in the MLIR context, can be mutable/mmaped/etc.
Resource arrays get encoded the same way as normal arrays, and on
decode we decide whether to use a resource or inline array based on
the size of the array data.

To simplify the decision making in the future of when to use
inline/out-of-line storage, a new API was added Support/Compiler.
This will help us stay consistent across the stack of when we
use which.

MODULAR_ORIG_COMMIT_REV_ID: d25a96f665ac0c2bc819f2a919b78dbae2ed62be