[KGEN] Enhance the elaborator to support basic regions. (#3174)
When resolving a kgen.call, we now notice regions on the calls and
squirrel them away in a side map, and use a new attribute to reason
about a signature parameter that is bound to a region.  When getting
to a call of that attribute, we subsitute the body of the region in
place of the call, unlike a call to a symbol which is recursively
elaborated out of line.

This also teaches elaboration of kgen.call_param to pass regions
to the formed kgen.call (instead of dropping them on the floor).

This is a nice step forward, but doesn't handle constraints or
parameters on the region yet.

MODULAR_ORIG_COMMIT_REV_ID: f7ea0bf3b476068882214280b70e5fea68c7d019