[KGEN] Generalize the Elaborator's opsToRewrite worklist, NFC. (#3179)
This extends the "opsToRewrite" worklist in the elaborator to
contain operations that need to be rewriten as well as special
commands.  The only command implemented so far is a "pop" operation
for the evaluator stack.  This allows management of scopes when
handling regions.

This isn't used yet, so it is NFC.

MODULAR_ORIG_COMMIT_REV_ID: c586a958fc98fbb61c89e067c6e8e89aa4ea8701