[KGEN] Teach the elaborator to handle input parameters to regions. (#3181)
When cloning a region body into the result function we now add the
operations in that region that need rewriting to the worklist of
the enclosing parameter rewriter.

The parameters that need to be rewritten are in a different scope
from the caller, so we push a scope on the evaluators list and
put our bindings into it.  This ensures that they get handled
correctly, even in the face of complicated nested resolution
cases.

Note: We still do not handle output parameters or constraints.
MODULAR_ORIG_COMMIT_REV_ID: 1a3f7ed2075fbcc01fef39887e1e582370fcdcfd