[KGEN] Implement support for returned parameters from regions on calls. (#3184)
This expands out the previous `PopEvaluatorStackType` marker in the
command list to a more rich `RegionReturn` entry that carries the
parameter expressions that need substitution and the declarations to
satisfy.  This allows the "evaluator stack pop" to properly evaluate
the return expressions in the callee stack, then set the decls in the
caller stack frame.

I believe that this finishes support for regions on call/call_param
notes.  This fixes Issue #2883.

MODULAR_ORIG_COMMIT_REV_ID: c945023bfa952b3116f3f969b394cb629f47840a