[KGEN] Eliminate the ? option for !pop.pointer. (#3188)
We can already form pointers to any type, including scalars which
can be any dtype.  This means we have two encodings for "void*":
pointer<?> and pointer<scalar<invalid>>.  We can't remove the later
encoding, so embrace it.

This has the knock-on effect of simplifying the type system by
eliminating a place that a null could occur, eliminating bugs and
simplifying the implementation of PointerType.

This fixes #2835.

MODULAR_ORIG_COMMIT_REV_ID: 061aec12163d7cf9fc2d28a1e7f4053ccbbcba57