[BEF/TFLite] Replace DenseElementsAttr with DenseArrayAttr (#3243)
This replaces all uses of DenseElements within BEF/TFLite with
DenseArrayAttr, with one small hiccup: `splat_array`. It is impractical
to write tests that contain shape `512x512` dense arrays (even as
resources because they are also quite long). To that end, I added a
special syntax for `m/tfl/p.constant` that parses and prints
`splat_array` and a single element and replicates it across the size of
the array.

Fixes [Internal link]

DenseElementsAttr needs to be replaced in KGEN but I'm going to make
that separate issue because this is killing my sanity.

Co-authored-by: abdul dakkak <adakkak@modular.com>
MODULAR_ORIG_COMMIT_REV_ID: 60824ea4e304df98430f79931518fd1f9f4cf55b