[Support] Copy over llvm TimeProfiler and adjust to support x-thread measurements (#3064)
[This change started in https://reviews.llvm.org/D133083, but the switch
to the high precision clock for cross-thread consistency is likely to be
contentious given I don't know if it works on Windows. Bringing into the
modular codebase while we iron this out. I suspect we'll end up redoing
all the profiling support anyway in the long term.]

This change allows a TimeProfiler profiling entry to be created on one
thread and finished on another. This was triggered by a use case where
we needed to trace the scheduling delay for work items ultimately
handled by independent worker threads.

The clock is changed to std::chrono::high_resolution_clock, both for its
increased precision (our work items can be in the us range), but also
for its cross-thread consistency (all threads see the same clock).

As a pleasant side effect it is now possible to pay for the cost of
constructing a profiling entry with a possibly expensive detail string
outside of the critical section we wish to trace.

MODULAR_ORIG_COMMIT_REV_ID: 8e8be53babd68f685ff8fc3f076fa9fd0fc44513