[Faux] Introduce `faux.verify.accuracy` (#3248)
The faux.verify.accuracy op takes two input tensors and verifies that
they are
within the specified tolerances.

The `faux.verify.accuracy` is used to verify that the `lhs` and `rhs`
tensors from two operations agree up to a specified tollerance. If the
`lhs` and `rhs` are within tollerance, then the result is `true`,
otherwise
the result is `false`.

The `lhs` and `rhs` must be of the same type and shape.

An optional `absoluteTolerance` and `relativeTolerance` can be
specified.
If not specified, the default absolute tollerance is 10^-6 and the
default
relative tollerance is 10^-8.

An optional `describeMismatch` flag can be specified. If specified, a
description of the mismatched values will be printed to stdout.

An optional `isNaNSensitive` flag can be specified. If specified, then
nans
are considered equal.

Usage:

```mlir
// Verify that the two tensors are within the default tollerances.
%ch1, %res = faux.verify.accuracy(%ch, %lhs, %rhs) :
  (!bef.chain, tensor<10xf32>, tensor<10xf32>) -> (!bef.chain, i1)

// Verify that the two tensors are within the specified absoluteTollerance.
%ch1, %res = faux.verify.accuracy(%ch, %lhs, %rhs) {
  absoluteTolerance = 10e-6
} : (!bef.chain, tensor<10xf32>, tensor<10xf32>) -> (!bef.chain, i1)

// Verify that the two tensors are within the specified absoluteTollerance
// and relativeTolerance.
%ch1, %res = faux.verify.accuracy(%ch, %lhs, %rhs) {
  absoluteTolerance = 10e-6, relativeTolerance = 10e-9
} : (!bef.chain, tensor<10xf32>, tensor<10xf32>) -> (!bef.chain, i1)

// Verify that the two tensors are within the default tollerances. In the
// event of a mismatch, print the values that mismatched as well as their
// indices.
%ch1, %res = faux.verify.accuracy(%ch, %lhs, %rhs) {
  describeMismatch = true
} : (!bef.chain, tensor<10xf32>, tensor<10xf32>) -> (!bef.chain, i1)

// Verify that the two tensors are within the default tollerances, but NaNs
// are considered equal.
%ch1, %res = faux.verify.accuracy(%ch, %lhs, %rhs) {
  isNaNSensitive = false
} : (!bef.chain, tensor<10xf32>, tensor<10xf32>) -> (!bef.chain, i1)
```

MODULAR_ORIG_COMMIT_REV_ID: a56ee8fd9fd1d707700f83d0bee8a1b47e99b2d5