[KGEN] Use a trick from River to get rid of getBodyBlock. NFC. (#3276)
Unfortunately, FunctionOpInterfaces is squating on the getBody()
method name, so we have to use a bit of a wild and crazy using
decl to get access to getBody(), but this is still a step forward.

I will file an MLIR issue to fix this upstream.

MODULAR_ORIG_COMMIT_REV_ID: f9eafc998ef88aff6e4777ba7dd101df1b5baf95