[KGEN] Add machinery for a new ParseLit library. (#3282)
This adds a new library with (the stub of) an MLIR translation that
parses a .lit file into an MLIR module.  This is a noop at the
moment, but this PR brings in the machinery needed to wire it up
and adds a testcase.

MODULAR_ORIG_COMMIT_REV_ID: 1012709cad58906b5dc5ef55331e478def865496