[Profiling] Move profiling setup/teardown into Runtime (#3278)
I sent Taewook on a bit of a wild goose chase trying to support
profiling third-party setup costs. In retrospect keeping the
setup/teardown isolated to runtime is much simpler. Though it would be
nice to be able to profile some of the stock-*-executor model setup
time, which is evaluated outside of the setupMLIR* call, the low-brow
approach is to just manually time that if it is important. With this we
can retire the TimeProfiler RAII class. Sorry about that Taewook!

MODULAR_ORIG_COMMIT_REV_ID: 3ec15de1462ce3d0b41cde1aa2cc2a9cc91f51e4