[Support] Add `!M.array` type. (#3357)
This PR adds the `M` dialect, a "builtin"-style dialect for Modular
code. Ideally, the contents of this dialect could be upstreamed if
necessary/desired (i.e., the dialect depends only on `MLIRIR`).

And this PR adds the `ArrayType`. This is a shaped type whose rank is
always 1 and is never dynamic. It can be used as types of elements
attributes. This type is useful throughout the Modular codebase where we
use ElementsAttr but restrict them to only 1 dimension.

I plan to replace our uses of `RankedTensorType` when we really just
want an array with this.

MODULAR_ORIG_COMMIT_REV_ID: 929212e9708fdc00c942100253d0c3bbce424e38