[LLCL] Add a NonAtomicallyReferenceCounted class, NFC. (#3433)
This class is like ReferenceCounted but does not use an std::atomic
for the base class.  It can be used for classes that aren't in a
threaded environment. RCRef works with it as expected.

A client is coming to test it.

MODULAR_ORIG_COMMIT_REV_ID: c0fce70ed6e477e9073b7c8314a942141dc96c8d