[KGEN] Rename hlkgen dialect to lit dialect. (#3480)
Now that we have a bit more clarity on the language direction, it is
clear that "hlkgen" is part of the Lightning⚡️ language lowering path
and desugars into the kgen dialect and friends.  Rename it to a more
cooperative name than :hulk: gen.

This fixes #3436

MODULAR_ORIG_COMMIT_REV_ID: 1d4d8662b90099ec2718c23e3940568b2d0bce53