[MDialect] Do not assume that APSInt has same bytesize as APInt (#3493)
This is not true in general (e.g. on
[Internal link]
This will perform the conversion via an extra vector allocation, but is
a lot more safe and portable across compilers/archs.

This fixes this issue for the Windows build

MODULAR_ORIG_COMMIT_REV_ID: 291a2db2792c7bfb91c4c07116746e2b5c32dbd1