[Lit] Parse value parameter declarations, add them to lit.func. (#3513)
This extends the parser to parse the value parameter
declarations in a def signature, capture their names but
dropping types/default args on the ground for now.

This is captured into the IR both as BBargs in the entry region
but also as a new valueParamNames array.  This is round tripped
correctly through the IR.

MODULAR_ORIG_COMMIT_REV_ID: e997637521a502d96128b2a7d38c2640bb518eaf