[KGEN] Introduce a variant type (#3497)
To the POP dialect. This is a structural variant type that reprensents a
value that is one of a list of types. Introduce associated ops:
`create`, `test`, and `bitcast` for interacting with the type.

The type and its ops were added to the POP dialect because it is a
natural home for "not exactly first-class language feature" stuff, such
as `kgen.struct.decl` vs `pop.struct`.

MODULAR_ORIG_COMMIT_REV_ID: 4f83daafaee6edb35629cf193e16d4781883d5ce