[Support] Add General MicroBenchmarking Infra (#3520)
This defines the common MicroBenchmark infrastructure used throughout
the Modular codebase. Code such as `grt.benchmark` and search in KGEN
will utilize this implementation (in subsequent PRs) to perform
microbenchmarking.

The implementation is designed to be used by any C++ code with no
dependency on LLCL or GraphRT. It is also designed to
be flexible and extensible, so that it can apply to many use cases.

Usage:

```cpp
 MicroBenchmark bench("name of my benchmark", [](MicroBenchmark::State &state) {
    for (auto _ : state) {
      <<< Do some work that needs to be benchmarked >>>
    }
 });
 MicroBenchmark::RunOptions runOptions;
 // You can optionally specify a prologue function. The prologue function
 // usually performs prep work that is needed by the main benchmark loop
 // body.
 runOptions.prologueFunction =  [](MicroBenchmark::State &state) {
    <<< Do some unmeasured prep work that needs to be done before the
    benchmark >>>
 };
 // Similarly an epilogue function can be specified. The epilogue function
 // usually performs cleanup work that is needed after the main benchmark
 // loop body.
 runOptions.epilogueFunction =  [](MicroBenchmark::State &state) {
    <<< Do some unmeasured cleanup work that needs to be done after the
    benchmark >>>
 };
 // You can then run the function with the specified options. The run
 // function returns an error if the benchmark failed to run.
 ErrorOrSuccess runResult = bench.run(runOptions);
 // You can then print the results of the benchmark.
 bench.report(llvm::outs(), ReportOptions());
```

Futures PRs will use this for
[Internal link] and
[KGEN [Internal link]

This resolves part of #3082

MODULAR_ORIG_COMMIT_REV_ID: f2374a989e4f08d5d600feb52202162545ad37fc