[KGEN] Add zap.debug_assert (#3388)
Asserts that the boolean value is true. If the value is false, the
optional message is printed to stderr and the program aborts.

Note: Unlike the `assert` operation in C, this operation is fully
removed if assertions are disabled. Therefore, production code should
not use this operation. This operation is meant for debugging and
verification purposes only.

Example:

```mlir
zap.debug_assert %cond, "a description of the condition"
```

example printout is

```
/Users/abduld/code/modular/KGEN/test/kgen/assert.mlir:6:kgen.func: failed assertion (zap.assert %cst, "assert failure!!!" : !pop.scalar<bool>) 'assert failure!!!'
```

MODULAR_ORIG_COMMIT_REV_ID: c441470e08fac6fd7ab7e6a13fd77e57abf4fe80