[KGEN] Add a simple variant optimization pass (#3541)
`prune-impossible-variants` does two things: it runs an analysis that
determines "known variant types" of variant values and propagates this
to `pop.variant.is` ops and function return types. `pop.variant.is` of a
variant that is known to be a particular type is folded to a constant
and functions that return a variant of a single known type is collapsed
to that type.

Co-authored-by: River Riddle <riddleriver@gmail.com>
MODULAR_ORIG_COMMIT_REV_ID: 5015b04db1cc431cb82ab63b2ce3b9fdbd3b59af