[GenericML] Refcounted Buffer class. (#3581)
This PR introduces refcounted `Buffer` class and `BufferHandle`, a RCRef
around Buffer. Buffer can be created from the pre-allocated object. In
such case, the object is not destroyed when the buffer is destroyed. Or
the underlying object can be allocated when the buffer is created, and
if that's the case the object is destroyed when the buffer is destroyed.

MODULAR_ORIG_COMMIT_REV_ID: 089a0acf536b9122ce04661bd05ec9b740fba195