[KGEN] Do not print the debug benchmarking warning in KGEN (#3622)
The micro-bench infra prints a warning when run in debug mode. This is
great when you are using it as standalone mode and there is something
actionable you can do, but it is not when you are in kgen. This changes
the micro-bench infra to take an option to allow the user to disable the
warning.

This also adjusts the runtime of micro-benchmarks in kgen to 20ms in
debug mode. Debug builds are not expected to generate performant code.
Ideally the timing is something that's passed by the user and is
controlled by CMake.

MODULAR_ORIG_COMMIT_REV_ID: f165b6b524e8c27b858ace6f38dff0c2f3a65bac