[Support] Add a trimmed mean summary statistic (#3623)
From [Wikipedia](https://en.wikipedia.org/wiki/Truncated_mean): "A
truncated mean or trimmed mean is a statistical measure of central
tendency, much like the mean and median. It involves the calculation of
the mean after discarding given parts of a probability distribution or
sample at the high and low end, and typically discarding an equal amount
of both. This number of points to be discarded is usually given as a
percentage of the total number of points, but may also be given as a
fixed number of points."

This will be used by kgen microbenchmarking to not favor one
implementation over another due to outliers

MODULAR_ORIG_COMMIT_REV_ID: ccae028d8765121396cb3803a21b4057fdaf3377