[Support] Add better discrimination between policy and implementation for BlobCache. (#3650)
This patch adds a CacheFindResult to the BlobCache which allows the
client to discriminate between the "error" state and the "not in cache"
state. It also provides an easy way to get a MemoryBufferRef when in the
value state, because that's the most common usage of the value when it
has a value. It forces the user to 'take' the value if they want to do
anything else with it such as returning the buffer from a function.

Fixes #3398

MODULAR_ORIG_COMMIT_REV_ID: 9b897532a7440da40c5fc6191974e8c2f405ae9f