[KGEN] Simplify code for shape indexing (#3677)
This simplifies the code for functions that operates on tensor shapes.
It also changes the code so that if a shape dimension is a known
constant, then that constant is used rather the shape being queried.

MODULAR_ORIG_COMMIT_REV_ID: 7d29c8d7ddd1e9e15248bb13f35fec7c9ccb5e31