[KGEN] Get debug symbols back in KGEN. (#3696)
The macOS linker ld64 does not copy dwarf information to output binaries
because it relies on dsymutil. Luckily for us, we usually have the
unified LLVM IR and therefore a single object file, so we can use it
directly when it's possible to do so - this recovers the dwarf
information for debugging.

This is absolutely a temporary and undesirable workaround until we
handle the linking manually.

MODULAR_ORIG_COMMIT_REV_ID: 80ad14ecd78d543fc07e2bdb4926bbaa9f4f2484