[KGEN] Add pop.memcpy op (#3738)
The `pop.memcpy` operation copies memory from the src operand to the
destination operand. The the source location to the destination location
must either be equal or non-overlapping. It copies `size` bytes of
memory
over. The `pop.memcpy` does not return a value.

The first argument is a pointer to the destination, the second is a
pointer
to the source. The third argument is an integer argument specifying the
number of bytes to copy.

If `size` is 0, it is no-op modulo the behavior of attributes attached
to
the arguments. If `size` is not a well-defined value, the behavior is
undefined. If `size` is not zero, both `dest` and `src` should be
well-defined, otherwise the behavior is undefined.

The `pop.memcpy` takes optional modifiers to either inline or mark the
memcpy as a volatile access. Those are specified via the `inline` and
`volatile` tags respectively.

Example:

```mlir
// Copy size bytes from src to dest.
pop.memcpy %dest, %src, %size : !pop.pointer<!pop.scalar<type>> to
                                !pop.pointer<!pop.scalar<i32>>

// Generate an inlined memcpy which copies size bytes from src to dest.
pop.memcpy inline %dest, %src, %size : !pop.pointer<!pop.scalar<i32>> to
                                        !pop.pointer<!pop.scalar<type>>

// Generate an inlined memcpy which copies size bytes from src to dest and
// mark the memcpy is a volatile access.
pop.memcpy inline volatile %dest, %src, %size :
                                        !pop.pointer<!pop.scalar<i32>> to
                                        !pop.pointer<!pop.scalar<i32>>
```

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: 517a83eba28f7240f30335139aebd9d38f207258