[Lit] Implement the first bit of name binding for type references. (#3784)
This implements the first part of the "second phases" of parsing:
name binding for types.  This is currently limited to var decls,
and is limited to direct references of named types (no support for
parameterized structs yet) but gets a lot of the first round of
plumbing in.

This works by teaching Scope about the declarations that contain
types that need to be name bound, and coming back to them after
all the top level declarations have been parsed and inserted into
the name scope tree.

When the scope has completed first phase parsing, each of these
entries are reparsed to extract their types, they have their
types resolved, and the resolved type is put back into the IR
(and/or diagnosed as an error).

MODULAR_ORIG_COMMIT_REV_ID: 40e602d4545759cf568a64f9500645f17d1db52e