[Lit] Split the expression parser out to its own file, NFC. (#3790)
This moves the expression parser out to its own .cpp file, shoving
a bunch of implementation logic into a self defined place, isolated
from the rest of the parser with the pImpl idiom.

This is only a few hundred lines of code now, but will grow as we
support the full grammar.  This also ensures we have better
separation of concerns between it and the rest of the parser.

MODULAR_ORIG_COMMIT_REV_ID: 3bc20bce9432b94a927dd9e3b8440810743b4d25