[KGEN] Add add dtype for KGEN's DType (#3813)
This introduces a new class KGENDType that extends the Support::DType
with an address type. It uses the KGENDType throughout

I know this can cause ambiguity, since the dtype in KGEN land is a
KGENDType and not a DType. I did try to introduce a EltType and use it
throughout, but that touches a lot of code. See
[Internal link]

This a better solution to [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 79f3eb395c5e0ffffa9a2f5c6353a5eaf1582b5a