[KGEN] Add pop.simd.gather to the POP dialect (#3798)
The `pop.simd.gather` reads scalar values from a SIMD vector of memory
locations and gathers them into one vector. The memory locations are
provided in the vector of pointers `base` as addresses. The memory is
accessed according to the provided mask. The mask holds a bit for each
vector lane, and is used to prevent memory accesses to the masked-off
lanes. The masked-off lanes in the result vector are taken from the
corresponding lanes of the `passthru` operand.

The `pop.simd.gather` is designed for conditional reading of multiple
scalar values from arbitrary memory locations in a single IR operation.
It
is useful for targets that support vector masked gathers and allows
vectorizing basic blocks with data and control divergence. Other targets
may support this intrinsic differently, for example by lowering it into
a
sequence of scalar load operations. The semantics of this operation are
equivalent to a sequence of conditional scalar loads with subsequent
gathering all loaded values into a single vector. The mask restricts
memory
access to certain lanes and facilitates vectorization of predicated
basic
blocks.

Assuming that %base is a vector of pointers [p0, p1, p2, p3] and %mask
is a
vector of i1 values [true, false, true, false], and the passthrough
vector
is [d0, d1, d2, d3] then the following op:

```mlir
%0 = pop.simd.gather %base[%mask], %passthrough :
                                    !pop.simd<4, !pop.pointer<type>>,
                                    !pop.simd<4, i1>
```

is equivalent to the following sequence of scalar loads:

```mlir
%0 = load %p0 : !pop.scalar<type>
%1 = load %p2 : !pop.scalar<type>
%2 = pop.constant(0) : !pop.simd<4, type>
%3 = pop.simd.insertelement %0, %2[0] : !pop.simd<4, type>
%4 = pop.simd.insertelement %d0, %3[1] : !pop.simd<4, type>
%5 = pop.simd.insertelement %1, %4[2] : !pop.simd<4, type>
%6 = pop.simd.insertelement %d3, %5[2] : !pop.simd<4, type>
```

In general, for some vector of pointers %base, mask %mask, and
passthrough
%pass instructions of the form:

```mlir
%0 = pop.simd.gather %base[%mask], %passthrough :
                                    !pop.simd<N, !pop.pointer<type>>,
                                    !pop.simd<N, i1>
```

is equivalent to the following sequence of scalar loads in C++:

```cpp
for (int i = 0; i < N; i++)
  result[i] = mask[i] ? *base[i] : passthrough[i];
```

This PR does not add LLVM lowering for the `pop.simd.gather` op

Depends on [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 8402d685a075d4ff920e02e280fceb1ba276ef26