[KGEN] Add a DataLayoutInterface and associated parameter operators (#3815)
This PR adds a `DataLayoutInterface` and two parameter operators:
`get_sizeof` and `get_alignof`, analogous to the C `sizeof` and
`alignof` keywords. These operators return in bytes the size and
alignments of the types.

This PR also adds a `pointerSize` parameter to `TargetInfoAttr`, for
lack of a better place to put it.

MODULAR_ORIG_COMMIT_REV_ID: b03336d907a743d257770efabad6c462cc7e2edb