[MOP] Add DType, TensorShape and Tensor types (#3853)
This family of types are intended to be 1:1 with the Modular
M::DType, M::TensorShape and M::Tensor types.

The TensorShape type is split out from Tensor so that we can
express ops which unpack tensors to their shapes. The DType
will not initially be constructable by any op and so could be
folded into the Tensor type, but I felt it was conceptually
clearer to split it out.

Note the severe restrictions on the Tensor type to make
clear exactly the implementation.

Supporting ops will follow this PR, it was getting a bit big.

See also !kgen.dtype and !zap.ndbuffer. Once the dust has
settled we should get together on a large whiteboard and
cross-pollinate/reconcile.

MODULAR_ORIG_COMMIT_REV_ID: 708d30f7191873b30e3af8240a384035ba9616e2