[BEF] Teach BEF how to encode/decode the mlir::UnitAttr (#3896)
This teaches BEF to lean how to encode and decode the mlir::UnitAttr

We need this to allow for code which defines ops that have unit
attributes. E.g. the matmul op can have a transpose attribute on the A
or B matrices. One can specify these attributes as bool, but that's a
bit more verbose.

MODULAR_ORIG_COMMIT_REV_ID: 4815e9bea6c058db1d2313a7730ffe7216cfe36a