[Lit] Introduce support for quashing reference errors. (#3910)
This adds a TypeCheckErrorType and a hasReferenceError bit for each
declaration to start keeping track of declarations that are
structurally broken.  References to these declarations are then
implicitly treated as pre-diagnosed which silences bogus downstream
errors.

For example, if you have: `def f(a: Int, a: Float):` then you don't
want to treat 'a' as having type Int or Float, because users of it
may expect one or the other and the compiler can't know the right
thing.  Better to have the user fix the first error first.

MODULAR_ORIG_COMMIT_REV_ID: 3f32b6d8a071e4ab0e9a7739489634ada2dc03d2