[KGEN] Add pop.simd.scatter op to POP (#3805)
The `pop.simd.scatter` stores scalar values from a SIMD vector of memory
locations and scatters them into a vector of pointers. The memory locations are
provided in the vector of pointers `base` as addresses. The memory is stored
according to the provided mask. The mask holds a bit for each vector lane, and
is used to prevent memory accesses to the masked-off lanes.

The `value` operand is a vector value to be written to memory. The `base`
operand is a vector of pointers, pointing to where the value elements should be
stored. It has the same underlying type as the value operand. The `mask`
operand, mask, is a vector of boolean values. The types of the `mask` and the
`value` operand must have the same number of vector elements.

In general, for some vector %value, vector of pointers %base, and mask %mask
instructions of the form:

```mlir
%0 = pop.simd.scatter %value, %base[%mask] : !pop.simd<N, type>
```

is equivalent to the following in C++:

```cpp
for (int i = 0; i < N; i++)
  if (mask[i])
    base[i] = value[i];
```

MODULAR_ORIG_COMMIT_REV_ID: 7e1c647124554a36341e35125acac6ee4d2917af