[Lit] Change Scope entries to hold scopes instead of Decls. (#3915)
This revises the ScopeValue type to hold a Scope* for a Decl instead
of the operation, since the scope has a pointer to the operation
but the decl requires a hash table lookup.  It also renames this type
to NameEntry since that is what it is.

While here, revise the hash table in the Scope object to be keyed
by a StringAttr instead of a StringRef, shrinking its size and reusing
pre-hashed identifiers.

Finally, this moves a bunch of the scope addition logic to
DeclResolver, tidying up and consolidating the logic.

MODULAR_ORIG_COMMIT_REV_ID: eb9661408f94fa7fcd7c0d525da8641a33f6e955