[Lit] Make the location translation logic more consistent, NFC. (#3927)
This sinks translateLocation to SharedState so declresolver can
use it. This allows it to take SMLoc's, which reduces the number
of locations that need to be translated.

This also moves getAsValue from MLIRValueRep to EmitterState,
which nicely simplifies some code.

MODULAR_ORIG_COMMIT_REV_ID: 0fce1e6014e6db206763a23147639f1c978ddcea