[KGEN] Fix documentation of zap.debug_assert (#3936)
Fixes the description of the zap.deug_assert op. The operation is not
currently removed in release builds. This is because we do not currently
have a notion of debug vs release builds for kgen kernels

Eventually debug_assert will be implemented as a library and predicated
on a buildInfo constraint

MODULAR_ORIG_COMMIT_REV_ID: 5bc932db0f62cf0ba20ca2710be144642bc04506