[Lit] Refactor if/while parsing a bit, fix a bug with else outside while. (#3992)
This adds a shared emitExprAsCondition method that is shared between
if/while emission (and will likely go elsewhere eventually) to factor
the conversion to MLIR i1 logic.

This also fixes a bug where an else follows a while loop but which is
less indented because it is part of an outer if statement.

MODULAR_ORIG_COMMIT_REV_ID: d2328b58542a19a01aa4e3e526b11d41f8af229e