[Lit] emit variable declarations in the right order. (#4059)
Previously we would always emit var decls at the start of the
region - this was simple and made sure that they dominated the
uses.  This worked, but had a few suboptimalities.

1) It was aethetically displeasing for functions, because it doesn't
   match the order you'd expect.
2) It made CHECK lines incrementally more annoying because they
   didn't follow the order of the statements that formed them.
3) It emited the members of a struct decl backwards as well, which
   could have layout implications.  This pushed me over the edge.

If we're going to change this behavior, it is best to do so before
we have more test cases, so go ahead and do it.

While here, add a testcase showing var members in a struct using
parameters from that struct.

MODULAR_ORIG_COMMIT_REV_ID: b023d751971ee26a3682c1eda4e9040402317da1