[Lit] Split lit-errors into two files, remove split-input from one. (#4061)
This moves the lexer errors out to their own file since they are
conceptually different from the parser and semantic errors in the
rest of the file.

This removes -split-input-file from the lit-errors driver because
we have proper error recovery for semantic and parse errors, so we
don't need it at all.

MODULAR_ORIG_COMMIT_REV_ID: 66abbb667a56d88107004667ad70c21429af9900