[Lit] implement support for calls that return void correctly. (#4062)
Previously we'd return a null pointer as the result of the walk,
but this is incorrect - null signifies an error emitting the value.

The actual desired result is an empty tuple value, but it will be
some time before we support tuples.  In the meantime, just use a
lit-dialect 'VoidAttr' as the standin.  We cannot use the MLIR
UnitAttr because it isn't a TypedAttr.

MODULAR_ORIG_COMMIT_REV_ID: 9b766446cbc1cb1cc6e8e93e3d365a67ff3270f1