[KGEN] Expand SignatureType to allow use of contextual types. (#4066)
The types embedded within a SignatureType were previously treated
as "isolated" from surrounding context, only allowing use of input
parameters defined within the signature itself.

This works for many things, but is a limited form of abstraction:
it turns out that this requirement is only needed when the signature
itself has input parameters.  In the case it doesn't (which, helpfully
is already sugared in the printer with function type syntax), it is
perfectly fine for the signature to refer to parameters defined in the
context of the signature.

Relaxing this allows new forms of abstraction, and helps unblock some
stuff in the lit world by making references to parameterized functions
more compositional.

MODULAR_ORIG_COMMIT_REV_ID: 099649dc132b8ba5ad3fba369220c08e39ac0b86