[Lit] Fix expression parsing line continuation; add paren patterns. (#4134)
This enhances the expression parser to know when to continue parsing
on the next line: as with python, we parse extended expressions when
in a ()/[]/{} block.  Unlike Python, we use a more flexible
indentation based scheme to get rid of the \'s in all or most all
normal code.

This happened to be required to parse the simple change to allow
paren patterns, the most trivial expression pattern.

MODULAR_ORIG_COMMIT_REV_ID: d25fcb9e7ca5bc3c193771caa5c4f89e23ccb248