[Support][TimeProfiler] Fix a bug in time profiler. (#4212)
This PR fixes the issue reported by #4145. The upstream LLVM version of
TImeProfiler didn't have this issue because the events are always
maintained by the thread-local stack and the stack can't be empty when
the `end` is called. In our version, we enabled explicit passing of
entries to enable the cross-thread tracing, so the stack can be empty
when the `end` is called which introduced the issue reported.

Closes #4145
MODULAR_ORIG_COMMIT_REV_ID: 372fe89a174b29b9133b3ef666948a4fd5b0d116